+++ NEWS.md
@@ -1,6 +1,6 @@
 # News
 
-## 0.0.5.9005 (development version)
+## 0.0.5.9006 (development version)
 
 License:
 
@@ -23,6 +23,10 @@ automatically vary line widths by group. (#134 @grantmcdermott)
 - `tpar()` now accepts standard `par()` arguments in addition to the
 `tinyplot`-specific ones. This allows users to set or query graphical parameters
 via a single convenience function, instead having to invoke `tpar` and `par` separately.
+- Users can write plots to disk using the new `filename` (alias `file`)
+convenience argument. The appropriate device type is automatically determined by
+the file extension, but must be one of ".png", ".pdf", or ".svg". (#143
+@grantmcdermott)
 
 Bug fixes: